requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
git+https://github.com/openai/CLIP.git
pillow
tqdm
pandas
numpy
Flask
cors
gunicorn
easyocr 